# docs/swagger.json
{"swagger":"2.0","info":{"description":"This is a sample forum API server written in Go.","title":"Forum API","contact":{},"version":"1.0"},"host":"https://localhost:8082","basePath":"/","paths":{"/login":{"post":{"description":"Authenticate user credentials and set session cookie","consumes":["application/x-www-form-urlencoded"],"produces":["text/html"],"tags":["auth"],"summary":"Log in a user","parameters":[{"type":"string","description":"User email","name":"email","in":"formData","required":true},{"type":"string","description":"Password","name":"password","in":"formData","required":true},{"type":"string","description":"Admin flag ('on' if admin)","name":"admin","in":"formData"}],"responses":{"302":{"description":"Redirects to /admin_page or / on success"},"403":{"description":"Forbidden","schema":{"$ref":"#/definitions/models.ErrorResponse"}}}}},"/registration":{"post":{"description":"Handles user registration form submission and user creation","consumes":["application/x-www-form-urlencoded"],"produces":["text/html"],"tags":["auth"],"summary":"Register a new user","parameters":[{"type":"string","description":"First Name","name":"firstName","in":"formData","required":true},{"type":"string","description":"Second Name","name":"secondName","in":"formData","required":true},{"type":"string","description":"Username","name":"username","in":"formData","required":true},{"type":"string","description":"Email Address","name":"email","in":"formData","required":true},{"type":"string","description":"Password","name":"password","in":"formData","required":true},{"type":"string","description":"Admin flag ('on' if admin)","name":"admin","in":"formData"}],"responses":{"302":{"description":"Redirects to /login on success"},"405":{"description":"Method Not Allowed","schema":{"$ref":"#/definitions/models.ErrorResponse"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/models.ErrorResponse"}}}}}},"definitions":{"models.ErrorResponse":{"type":"object","properties":{"errorMessage":{"type":"string"},"errorNum":{"type":"integer"}}}}}